Store crypted keys in Session

+++ module/Application/src/Service/UserKeyService.php
@@ -8,6 +8,7 @@ use Application\Model\User;
 use Zend\Crypt\BlockCipher;
 use Zend\Db\TableGateway\TableGatewayInterface;
 use Zend\Math\Rand;
+use Zend\Session\Container;
 
 class UserKeyService
 {
@@ -22,6 +23,11 @@ class UserKeyService
      */
     protected $blockCipher;
 
+    /**
+     * @var Container
+     */
+    protected $container;
+
     public function __construct(TableGatewayInterface $keysTable, BlockCipher $blockCipher)
     {
         $blockCipher->setBinaryOutput(true);
@@ -33,23 +39,32 @@ class UserKeyService
     /**
      * Add key to DB and generate cookie value
      *
-     * @param string $key User's encryption key
+     * @param string $key      User's encryption key
      * @param User   $user
+     * @param bool   $remember Save to DB or not
      *
      * @return string
      */
-    public function saveUserKey($key, User $user)
+    public function saveUserKey($key, User $user, $remember = false)
     {
         $cookieKey = Rand::getString(20);
         $this->blockCipher->setKey($cookieKey);
+        $cryptedKey = $this->blockCipher->encrypt($key);
 
-        $this->keysTable->insert([
-            'user_id' => $user->getId(),
-            'key'     => $this->blockCipher->encrypt($key),
-            'date'    => date('Y-m-d H:i:s'),
-        ]);
+        If ($remember) {
+            $this->keysTable->insert([
+                'user_id' => $user->getId(),
+                'key'     => $cryptedKey,
+                'date'    => date('Y-m-d H:i:s'),
+            ]);
 
-        $id = $this->keysTable->getLastInsertValue();
+            $id = $this->keysTable->getLastInsertValue();
+        } else {
+            $id = 0;
+        }
+
+        $this->getContainer()->cryptedKey = $cryptedKey;
+
         return $id . "-" . $cookieKey;
     }
 
@@ -63,19 +78,28 @@ class UserKeyService
      */
     public function getUserKey($value, User $user)
     {
-        $data  = $this->parse($value);
-        $where = [
-            'id'      => $data['id'],
-            'user_id' => $user->getId(),
-        ];
+        $data      = $this->parse($value);
+        $container = $this->getContainer();
 
-        $entity = $this->keysTable->select($where)->current();
-        if (! ($entity instanceof EncryptionKey)) {
-            throw new InvalidUserKeyException('Key not found');
+        if ($container->offsetExists('cryptedKey')) {
+            $cryptedKey = $container->cryptedKey;
+        } else {
+            $where = [
+                'id'      => $data['id'],
+                'user_id' => $user->getId(),
+            ];
+
+            $entity = $this->keysTable->select($where)->current();
+            if (! ($entity instanceof EncryptionKey)) {
+                throw new InvalidUserKeyException('Key not found');
+            }
+
+            $cryptedKey = $entity->getKey();
         }
 
         $this->blockCipher->setKey($data['key']);
-        $key = $this->blockCipher->decrypt($entity->getKey());
+        $key = $this->blockCipher->decrypt($cryptedKey);
+
         if (! $key) {
             throw new InvalidUserKeyException('Decryption fail');
         }
@@ -111,10 +135,14 @@ class UserKeyService
     {
         $data = $this->parse($value);
 
-        $this->keysTable->delete([
-            'id'      => $data['id'],
-            'user_id' => $user->getId(),
-        ]);
+        if ($data['id']) {
+            $this->keysTable->delete([
+                'id'      => $data['id'],
+                'user_id' => $user->getId(),
+            ]);
+        }
+
+        $this->getContainer()->offsetUnset('cryptedKey');
 
         return $this;
     }
@@ -129,5 +157,29 @@ class UserKeyService
         return $this->keysTable->delete([
             '`date` < NOW() - INTERVAL 2 WEEK',
         ]);
+    }
+
+    /**
+     * @return Container
+     */
+    public function getContainer()
+    {
+        if (! $this->container) {
+            $this->setContainer(new Container('EncryptionKey'));
+        }
+
+        return $this->container;
+    }
+
+    /**
+     * @param Container $container
+     *
+     * @return self
+     */
+    public function setContainer($container)
+    {
+        $this->container = $container;
+
+        return $this;
     }
 }
